@@ -1,21 +1,22 @@
 {
-  "name": "portfolio",
-  "version": "0.1.0",
+  "name": "react-portfolio",
+  "version": "2.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
     "framer-motion": "^4.1.17",
+    "locomotive-scroll": "^4.1.4",
     "normalize.css": "^8.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-particles-js": "^3.5.3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.3.0",
-    "tsparticles": "^1.31.1",
-    "web-vitals": "^1.1.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.4",
+    "react-scripts": "^2.1.3",
+    "react-tsparticles": "^2.6.0",
+    "styled-components": "^5.3.6",
+    "tsparticles": "^2.6.0",
+    "web-vitals": "^3.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -40,9 +41,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "description": "",
-  "main": "index.js",
-  "author": "",
-  "license": "ISC"
+  }
 }
