@@ -1,20 +1,17 @@
+import { Routes, Route, useLocation } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
 import { lightTheme } from "./components/Themes";
-import { AnimatePresence } from "framer-motion";
+import { AnimatePresence } from "framer-motion/dist/framer-motion";
 import GlobalStyle from "./globalStyles";
+import React from "react";
 
 //Components
 import Main from "./components/Main";
 import AboutPage from "./components/AboutPage";
 import BlogPage from "./components/BlogPage";
 import WorkPage from "./components/WorkPage";
-import MySkillsPage from "./components/MySkillPage";
+import MySkillPage from "./components/MySkillPage";
 import SoundBar from "./subComponents/SoundBar";
-import {
-  Routes,
-  Route,
-  useLocation,
-} from "react-router-dom/cjs/react-router-dom.min";
 
 function App() {
   const location = useLocation();
@@ -41,7 +38,7 @@ function App() {
 
             <Route path="/work" element={<WorkPage />} />
 
-            <Route path="/skills" element={<MySkillsPage />} />
+            <Route path="/skills" element={<MySkillPage />} />
             {/* Below is to catch all the other routes and send the user to main component,
 you can add custom 404 component or message instead of Main component*/}
             <Route path="*" element={<Main />} />
